Fix Rust compilation errors for Tauri build
- Fix feed-rs Content enum: use MediaType::Html/Text instead of Content::Html/Text
- Fix into_filter (unstable): use into_iter().filter()
- Fix Duration::from_secs type mismatch: cast u32 to u64
- Add json feature to reqwest for blocking response.json()

+++ src-tauri/src/scheduler.rs
@@ -33,7 +33,7 @@ impl Scheduler {
                     s.map(|s| s.refresh_interval_hours).unwrap_or(6)
                 };
 
-                let interval_secs = interval * 3600;
+                let interval_secs = (interval as u64) * 3600;
 
                 // Generate on first run
                 {
